Move implementation details from hpp to cpp
