+++ kizhin.evgeniy/T2/io-utils.hpp
@@ -1,8 +1,13 @@
 #ifndef SPBSPU_LABS_2025_TP_A_KIZHIN_EVGENIY_T2_IO_UTILS_HPP
 #define SPBSPU_LABS_2025_TP_A_KIZHIN_EVGENIY_T2_IO_UTILS_HPP
 
-#include "input-utils.hpp"
-#include "output-utils.hpp"
+#include <iosfwd>
+#include "data-struct.hpp"
+
+namespace kizhin {
+  std::istream& operator>>(std::istream&, DataStruct&);
+  std::ostream& operator<<(std::ostream&, const DataStruct&);
+}
 
 #endif
